+++ NEWS.md
@@ -1,4 +1,5 @@
 * Adding support for rendering single MIDI chords
+* Adding support for rendering MIDI chord sequences
 
 # abcR 0.0.0.9002