+++ NEWS.md
@@ -1,3 +1,5 @@
+# abcR 0.0.0.9003
+
 * Adding support for rendering single MIDI chords
 * Adding support for rendering MIDI chord sequences